requirements.txt
beautifulsoup4==4.9.3
click==7.1.2
geopandas==0.8.1
joblib #==1.0.1
lxml==4.6.2
pandas==1.1.5
pyproj==2.6.1
pystac #==1.0.1
# rasterio==1.1.8
rio_cogeo #==2.3.1
ruamel.yaml.clib==0.2.0
ruamel.yaml==0.16.10
SQLAlchemy==1.3.2
tabulate==0.8.7
utm==0.7.0

--extra-index-url="https://packages.dea.ga.gov.au"
odc_apps_dc_tools # ==0.1.dev1151+g7a7947e
# 'odc-apps-cloud[GCP,THREDDS]'
# odc-ui
# odc-index
# odc-geom
# odc-algo
# odc-io
# odc-aws
# odc-aio
# odc-dscache
# odc-dtools
